shield360-django/requirements.txt
Django==3.2.25
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
djongo==1.3.6
pymongo[srv]>=4.6
python-dotenv>=1.0
whitenoise>=6.6
sqlparse==0.2.4
